feat: implement mosque search results handling and UI state conversion
