+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/feature/mosque/NearbyMosquesUiState.kt
@@ -1,5 +1,6 @@
 package net.thechance.mena.faith.presentation.feature.mosque
 
+import net.thechance.mena.faith.domain.entity.Mosque
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
 
@@ -30,40 +31,15 @@ internal data class Coordinate(
 )
 
 @OptIn(ExperimentalUuidApi::class)
-internal val fakeMosqueList = listOf(
-    MosqueUiState(
-        id = Uuid.parse("1e6f8a10-7dec-11d0-a765-00a0c91e6bf1"),
-        name = "Al-Fath Mosque",
-        imageUrl = "https://example.com/images/alfath.jpg",
-        distance = 0.8,
-        coordinate = Coordinate(30.0444, 31.2357)
-    ),
-    MosqueUiState(
-        id = Uuid.parse("1e6f8a11-7dec-11d0-a765-00a0c91e6bf2"),
-        name = "Al-Azhar Mosque",
-        imageUrl = "https://example.com/images/alazhar.jpg",
-        distance = 2.3,
-        coordinate = Coordinate(30.0459, 31.2625)
-    ),
-    MosqueUiState(
-        id = Uuid.parse("1e6f8a12-7dec-11d0-a765-00a0c91e6bf3"),
-        name = "Amr Ibn Al-As Mosque",
-        imageUrl = "https://example.com/images/amribnalas.jpg",
-        distance = 4.1,
-        coordinate = Coordinate(30.0061, 31.2315)
-    ),
-    MosqueUiState(
-        id = Uuid.parse("1e6f8a13-7dec-11d0-a765-00a0c91e6bf4"),
-        name = "Sultan Hassan Mosque",
-        imageUrl = "https://example.com/images/sultanhassan.jpg",
-        distance = 3.7,
-        coordinate = Coordinate(30.0299, 31.2593)
-    ),
-    MosqueUiState(
-        id = Uuid.parse("1e6f8a14-7dec-11d0-a765-00a0c91e6bf5"),
-        name = "Mohamed Ali Mosque",
-        imageUrl = "https://example.com/images/mohamedali.jpg",
-        distance = 5.5,
-        coordinate = Coordinate(30.0286, 31.2599)
+internal fun Mosque.toUiState(distance: Double): MosqueUiState {
+    return MosqueUiState(
+        id = id,
+        name = name,
+        imageUrl = imageUrl,
+        distance = distance,
+        coordinate = Coordinate(
+            latitude = coordinates.latitude,
+            longitude = coordinates.longitude
+        )
     )
-)
+}
